safari-534.53-branch/LayoutTests/fast/js/regexp-char-insensitive-expected.txt

This test checks the case-insensitive matching of character literals.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS /å/i.test('/å/') is true
PASS /å/i.test('/Å/') is true
PASS /Å/i.test('/å/') is true
PASS /Å/i.test('/Å/') is true
PASS /å/i.test('P') is false
PASS /å/i.test('PASS') is false
PASS /Å/i.test('P') is false
PASS /Å/i.test('PASS') is false
PASS 'PASS'.match(/Å/i) is null
PASS 'PASS'.match(/Å/i) is null
PASS 'PASå'.replace(/å/ig, 'S') is 'PASS'
PASS 'PASå'.replace(/Å/ig, 'S') is 'PASS'
PASS 'PASÅ'.replace(/å/ig, 'S') is 'PASS'
PASS 'PASÅ'.replace(/Å/ig, 'S') is 'PASS'
PASS 'PASS'.replace(/å/ig, '%C3%A5') is 'PASS'
PASS 'PASS'.replace(/Å/ig, '%C3%A5') is 'PASS'
PASS successfullyParsed is true

TEST COMPLETE
